Load locations based on Distance Calculator

+++ Server/Hambasafe.Api/Models/v1/CoordinateModel.cs
@@ -2,8 +2,8 @@
 {
     public class CoordinateModel
     {
-        public decimal? Latitude { get; set; }
-        public decimal? Longitude { get; set; }
-        public decimal? Distance { get; set; }
+        public double? Latitude { get; set; }
+        public double? Longitude { get; set; }
+        public double? Distance { get; set; }
     }
 }
